:truck: refactor: 优化组织 types 类型定义与目录
