+++ src/functions/createGlobalStyle.tsx
@@ -1,12 +1,11 @@
 import { Global } from '@emotion/react';
 import { serializeStyles } from '@emotion/serialize';
-import { Interpolation } from '@emotion/styled';
 import { memo } from 'react';
 
 import { useTheme } from '@/hooks';
-import { Theme } from '@/types';
+import { CSSStyle, Theme } from '@/types';
 
-interface GlobalTheme {
+export interface GlobalTheme {
   theme: Theme;
 }
 
@@ -14,9 +13,7 @@ interface GlobalTheme {
  * 创建全局样式
  * @param styles
  */
-export const createGlobalStyle = (
-  ...styles: Array<TemplateStringsArray | Interpolation<GlobalTheme>>
-) =>
+export const createGlobalStyle = (...styles: CSSStyle<GlobalTheme>) =>
   memo((props) => {
     const theme = useTheme();
     return <Global styles={serializeStyles(styles, undefined, { ...props, theme })} />;
