+++ src/types/appearance.ts
@@ -1,0 +1,3 @@
+export type DisplayTheme = 'dark' | 'light';
+
+export type ThemeMode = 'auto' | 'dark' | 'light';
