+++ src/types/css.ts
@@ -1,0 +1,4 @@
+import { Theme } from '@/types/theme';
+import { Interpolation } from '@emotion/styled';
+
+export type CSSStyle<T = Theme> = Array<TemplateStringsArray | Interpolation<T>>;
